refactor(agent): extract publish/share write-pipeline into mixin holder
Moves the publish/share write-pipeline (~2.5k lines) out of dkg-agent.ts
into PublishMethods (extends DKGAgentBase): gossip publish, publishAsync/
_publish, update/share/conditionalShare, assertionFinalize + precomputed
attestation, curated-key + inline-encrypt resolution, publishFrom{Finalized
Assertion,SharedMemory}, and the on-chain register/batch/identity entrypoints.
Assembled onto DKGAgent via applyMixins + interface merge.

- moved methods take `this: DKGAgent` and are public (cross-called from main)
- 13 sibling methods the cluster calls widened private->public
- two DKGAgent.<static> refs -> DKGAgentBase (SWM_SUBSTRATE_FANOUT_TIMEOUT_MS,
  SWM_GOSSIP_FAILURE_MAX_TRACKED_CGS)
- no stranded field decls (extractor now checks)

Pure relocation. dkg-agent.ts 17245 -> 14686. Typecheck clean; unit (123)
+ agent.test.ts (126) + gossip/publish suites (gossip-publish-handler 9,
swm-gossip-publish-failure 8, swm-gossip-signing 6) green.
